# backend-golang/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/auth/register_user": {
            "post": {
                "description": "Регистрация пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API для авторизации и регистрации пользователя"
                ],
                "summary": "Register user",
                "operationId": "register_user",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SignUpRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data",
                        "schema": {
                            "$ref": "#/definitions/models.SignInResponse"
                        }
                    }
                }
            }
        },
        "/auth/user_by_auth": {
            "post": {
                "description": "Авторизация пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API для авторизации и регистрации пользователя"
                ],
                "summary": "Auth user",
                "operationId": "user_by_auth",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SignInRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data",
                        "schema": {
                            "$ref": "#/definitions/models.SignInResponse"
                        }
                    }
                }
            }
        },
        "/chat/add_user_to_chat": {
            "post": {
                "description": "Добавить юзера в чат",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API для работы с чатами и сообщениями"
                ],
                "summary": "Add user to chat",
                "operationId": "add_user_to_chat",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddUserToChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        },
        "/chat/change_user_info": {
            "post": {
                "description": "Изменить информацию о юзере",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API для работы с чатами и сообщениями"
                ],
                "summary": "Change user info",
                "operationId": "change_user_info",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ChangeUserInfoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    }
                }
            }
        },
        "/chat/chats_by_token": {
            "post": {
                "description": "Чаты по токену авторизации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API для работы с чатами и сообщениями"
                ],
                "summary": "Get chats by auth token",
                "operationId": "chats_by_token",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ByTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data",
                        "schema": {
                            "$ref": "#/definitions/models.Chats"
                        }
                    }
                }
            }
        },
        "/chat/create_chat_return_users": {
            "post": {
                "description": "Создать чат",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API для работы с чатами и сообщениями"
                ],
                "summary": "Create chat",
                "operationId": "create_chat_return_users",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateChatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data",
                        "schema": {
                            "$ref": "#/definitions/models.ChatUsers"
                        }
                    }
                }
            }
        },
        "/chat/create_message": {
            "post": {
                "description": "Создать сообщение",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API для работы с чатами и сообщениями"
                ],
                "summary": "Create message",
                "operationId": "create_message",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data",
                        "schema": {
                            "$ref": "#/definitions/models.CreateMessageResponse"
                        }
                    }
                }
            }
        },
        "/chat/messages_by_chat_id": {
            "post": {
                "description": "Сообщения по ИД чата",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API для работы с чатами и сообщениями"
                ],
                "summary": "Get messages by chat_id",
                "operationId": "messages_by_chat_id",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MessagesByChatIdRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data",
                        "schema": {
                            "$ref": "#/definitions/models.Messages"
                        }
                    }
                }
            }
        },
        "/chat/users_by_chat_id": {
            "post": {
                "description": "Юзеры по ИД чата",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API для работы с чатами и сообщениями"
                ],
                "summary": "Get users by chat_id",
                "operationId": "users_by_chat_id",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ByChatIdRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data",
                        "schema": {
                            "$ref": "#/definitions/models.Users"
                        }
                    }
                }
            }
        },
        "/chat/users_by_name": {
            "post": {
                "description": "Юзеры по имени",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "API для работы с чатами и сообщениями"
                ],
                "summary": "Get users by name",
                "operationId": "users_by_name",
                "parameters": [
                    {
                        "description": "credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UsersByNameRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data",
                        "schema": {
                            "$ref": "#/definitions/models.Users"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.AddUserToChatRequest": {
            "type": "object",
            "properties": {
                "auth_token": {
                    "type": "string"
                },
                "chat_id": {
                    "type": "string"
                },
                "user_link": {
                    "type": "string"
                }
            }
        },
...
        "models.ByTokenRequest": {
            "type": "object",
            "properties": {
                "auth_token": {
                    "type": "string"
                }
            }
        },
        "models.ChangeUserInfoRequest": {
            "type": "object",
            "properties": {
                "auth_token": {
                    "type": "string"
                },
                "new_about": {
                    "type": "string"
                },
                "new_image": {
                    "type": "string"
                },
                "new_link": {
                    "type": "string"
                },
                "new_name": {
                    "type": "string"
                }
            }
        },
...
        "models.ChatUsers": {
            "type": "object",
            "properties": {
                "chat_id": {
                    "type": "string"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.User"
                    }
                }
            }
        },
        "models.Chats": {
            "type": "object",
            "properties": {
                "chats": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Chat"
                    }
                }
            }
        },
        "models.CreateChatRequest": {
            "type": "object",
            "properties": {
                "auth_token": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "users_links": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.CreateMessageRequest": {
            "type": "object",
            "properties": {
                "auth_token": {
                    "type": "string"
                },
                "chat_id": {
                    "type": "string"
                },
                "reply_msg_id": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                }
            }
        },
        "models.CreateMessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "$ref": "#/definitions/models.Message"
                },
                "reply_msg": {
                    "$ref": "#/definitions/models.Message"
                }
            }
        },
...
        "models.Messages": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Message"
                    }
                }
            }
        },
...
        "models.SignInRequest": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "models.SignInResponse": {
            "type": "object",
            "properties": {
                "about": {
                    "type": "string"
                },
                "auth_token": {
                    "type": "string"
                },
                "image_id": {
                    "type": "string"
                },
                "last_online": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.SignUpRequest": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "about": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_id": {
                    "type": "string"
                },
                "last_online": {
                    "type": "string"
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    }
}